+++ HISTORY.md
@@ -1,7 +1,8 @@
 ## Changelog ##
 
-### Unreleased (Dev time: 1.0 hours) ###
+### Unreleased (Dev time: 2.0 hours) ###
 - Tweak: Improve performance of DB queries to prevent timeouts. (Issue #93)
+- Tweak: Add details about different post types in system info. (Issue #100)
 
 ### 2015-02-14 - v5.4.1 - (Dev time: 0.5 hours) ###
 - Tweak: Use Google CDN for jQuery UI CSS, instead of ASP.NET